Adds a basic prototype with file upload
Currently supplies an /upload route (accessible from the home page)
which accepts an xlsx file and then redirects to a success page.  The
target for the file is configured in the app.json although we should
probably move to env vars in future.

This is intended as a starting point from which we can implement better
abstractions to handle the upload.