First pass of intended user journey
This adds a documentation page that shows the various steps we are
aiming to support in the user journey for data import. We are quite
early on this – there are some key ideas in here from `slang` such as
users needing to select ranges and map columns, but the minutiae of
how that is done is yet to be defined.

But the general principle is here – that this is not just for demo
purposes, but should evolve into an explanation of how the user
journey works and how the components should be used.

This should be kept up to date as the user journey evolves.

The page doesn't use images – instead it uses the real GOV.UK design
system components but scaled down to fit into a little card. This
should be easier for us to iterate as we move the journey around,
rather than having to open an image editor every time.

To see the page in full glory, use e.g.

    python3 -m http.server