Adds a basic prototype with file upload (#2)
Currently supplies an /upload route (accessible from the home page) which accepts an xlsx file and then redirects to a success page. The target for the file is configured in the app.json although we should probably move to env vars in future.

The importer library is initialised which will configure the router to specify the routes/controllers used for upload. Currently templates still live in the prototype, but can be installed in the prototype from the manage prototype page (as with any other prototype kit plugin).

This is intended as a starting point from which we can implement better abstractions to handle the upload.