Ensure lib importer dependencies are installed for new prototype
When creating a prototype, we reference the local importer and as a
result it will use its own local node_modules, if it is not up to date
then this will introduce problems in the prototype.

Now we do an `npm i` in lib/importer to make sure its dependencies are
up to date.

Fixes #8