Refactor endpoints (#14)
Refactors the endpoints to change the behaviour of the system. Before this PR each endpoint handled the GET and POST requests for each step, rendering the template, processing the state, and then redirecting to the next step. Although this worked, it did not empower users of the plugin to use it how they want to.

After this PR, the plugin is composed of:

    Macros that will render specific parts of the upload state. e.g. select a sheet, or review the data etc.
    Template functions that can be used in forms to specify how to process the current step, and then the URL to redirect to afterwards
    Endpoints to handle POST requests from forms and then redirect to the requested URL.
    Templates showing how the macros can be used either for education, or to be copied into the users' prototype.