Auto create tmp directory if one not specified
Currently the user can specify an uploadPath in the prototype's
config.json file (with a key of uploadPath) and it needs to be a
directory that exists.  If no uploadPath is present in the config, then
the library will instead create a new temp directory to use for that
run.