Macro for table assets (#21)
When using a scoped package (@register-dynamics/importer) the path to
load the assets is no longer plugin-assets/importer but now includes the
scoped package name.

To resolve this we have introduced a new macro for loading the table
assets directly, without the user needing to know what they are.  This
might be prettier if we had an obvious way to get the name of the
package the code lives in.  We might in future load the package.json and
read/encode and use it.