First pass at documentation included with plugin
This commit contains more documentation set out in the style of a gov
Design System, splitting things in components and patterns, and
including things about when to use the pattern or not. It also
includes installation instructions for configuring and using the tool.

The idea here is that we can add more functionality when we are inside
the Prototype Kit – we can for instance link to the correct template
creation page as part of the user journey. But we may want this
documentation to also be available as a static site for perusal by
people who haven't installed the plugin yet. For this purpose we can
now tag elements with `data-plugin-only` or `data-static-only` to
control which elements are displayed where – in some future process,
we can remove the data-plugin-only elements for static site usage.