Build and deploy plugin documentation as a static site
The general approach here is to run the Prototype Kit with the plugin
loaded, get the documentation pages that we know exist, work out what
assets they all need, and then request those too. We do it this way
because the documentation can reuse assets from the Prototype Kit
which are not necessarily even on disk (e.g. some CSS is rendered from
Sass) so the cleanest way to get those assets is just to download them

This requires a copy of `hq` which allows us to both remove the plugin
only elements and also get the assets we need from HTML attributes.