Do not remove visual selection when focus is lost (#28)
Currently when the focus is removed from the table, the current visual selection is also removed. Whilst we will have taken what we want from the table at this point (the cell range) it may confuse users to see the selection disappear as they click a nearby submit button.

To configure this behaviour (remember the selection) the table should contain a data attribute called persist-selection like

<table class="selectable" data-persist-selection="true">

Any other value, or the absence of the attribute will result in the old behavious of the selection being removed when the table focus is lost.