Select header rows from uploaded file (#23)
Adds a new header selection screen to the flow, where users can select which cells contain the headers that they are interested in. This selection is then used as the definition of the headers, and will be used to:

    Map the columns to output fields
    Define the horizontal/vertical bounds of the data rows.

The new addition to the table uses specifically named fields (we might in future use html data attributes) to store the bounds of the selection, allowing hidden fields to provide the form data to the backend. The newly selected header range is stored in the frontend session and then passed to the backend as required (along with a translation from header range to data range).