Show an error if an empty sheet is selected
Although we should show all of the sheets available in the sheet
selector (to avoid confusion) we may end up in a situation where a user
selects an empty sheet.  Previously, we just crashed with a bad range
when this happened, but now we instead show the user an error explaining
that they need to choose a non-empty sheet.