Add a footer selection step (#54)
Adds an optional step after header selection for selecting any footer cells in the current sheet. To do this the PR contains a refactor of the range selector so that users should now use importerHeaderSelector or importerFooterSelector rather than the lower-level importerRangeSelector; this allows us to fetch specific rows for each selector type in the outer macro, passing the rows to the inner importerRangeSelector macro.

The new footer selection step allows for no selection, either by pressing the Skip button or by not selecting any cells in the presented table.