Automatically install the plugin router
Here we use a postinstall hook to automatically add the required
router lines to the app/routes.js, if it exists. We check first that
this looks like a Prototype Kit prototype and not some other app.

Unfortunately we can't automatically reverse this process because npm
doesn't run the pre/postuninstall hooks of dependent packages, because
it sucks.