Add playwright for running UI tests (#66)
Adds playwright to the basic prototype so that we can test the prototype where components are built and tested. Currently only tests the happy path using tribbles.xlsx, but provides the foundation for further testing. Does not yet test new prototype creation, this will be done as a separate PR