Fix off by one error in small files (#88)
Currently there's an off-by-one error when displaying a small file with fewer rows than the number used in previews (10). This fixes the off-by-one and introduces a test to ensure we don't re-introduce it.

This should resolve #84 but doesn't resolve the issue where if you select headers and footers that leave no data to be processed, we just show an error on the mapping page.

This currently requires only a single worker for running the tests to avoid issues with session state.