Update makefile to simplify updates
To make it easier to update dependencies when dependabot submits a new
update, this adds a new target `update-deps` which will update the importer and the
development prototype.

To make this work, had to restructure the `prototype` target so that it
doesn't complain for other targets when `NAME=` is missing.