Show upload preview without error
Adds a new macro which will show the upload form without relying on any
macros that are not installed because we are showing the preview by just
showing the template on the page.

As a result the preview.show() macro needs to import all of the govuk
components (which are not installed in the importer, but will be when
showing the preview) for any govuk components.

Unfortunately, when installing this template, it will keep the block
that decides to show the preview and will likely generate support
questions around its presence.