Re-introduce ODS file support in tests
Currently the tests will fail with ODS files as the number format bug in
the xlsx lib's support for ODS will call console.error() and in tests,
this causes the test to fail.

To resolve this we disable that functionality in backend tests (until we
can be more explicit for only ODS files) and wrap the tests in each file
in a describe() function to keep the changes isolated.