Enable loading of views from the plugin
This commit allows the importer plugin to provide template files for
rendering in a local views folder, with only access to the express
router.

Adding a new folder, without making it accessible to end-users would
typically require a much earlier step allowing the plugin to provide
view folders for its own use, but that is not available.  This commit
instead adds a middleware function that intercepts all requests to
/importer/* and checks if the local views folder is already included in
the list of view folders.

The introspection of the underlying nunjucks configuration, and
modification of the file loader that is current installed relies on a
reference to the express app running the prototype, which is the reason
for the middleware as the app is accessible through each request.