Change the structure for mapping fields in config
Changes configuration so that instead of just having a field name, it
now accepts a field type (string or number) and a boolean which suggests
whether the field is required in the mapping or not.

The codebase still supports the old style configuration (fields are a
list of names) but defaults to not-required and string fields.