Change the structure for mapping fields in config (#165)
* Change the structure for mapping fields in config

Changes configuration so that instead of just having a field name, it
now accepts a field type (string or number) and a boolean which suggests
whether the field is required in the mapping or not.

The codebase still supports the old style configuration (fields are a
list of names) but defaults to not-required and string fields.

* Update documentation with link to config tool

* Update mappings to use new structure

* More testing of avg/sum in review

* Update config to use new structure

* Show type and required flag on config page

* Use "text" instead of "string" for the config type

* Specify type and optionality when adding field

* Shorten wait to save config