Publish package on push (#201)
When merging to main, publish the lib/importer as a package to https://www.npmjs.com/package/@register-dynamics/importer

This github action only runs when a github release is published, so this PR does not involve itself in updating the package version in package.json.

If in future we decide to publish multiple packages from here we can add a task to check both package.json have been updated and keep the versions in sync.