Apply validation functions to mapping. (#192)
Use the attribute-types to ensure that rows are validated according to the type and required fields when performing the validation. This should generate errors and warnings for the review page.

In addition to these changes, the warning/errors are defined such that validation failures are errors, but structural errors (such as missing rows, or mismatched row lengths) are recorded as warnings. For errors, it is now possible to request the row data in code for any errors, which in future will allow the review page to show exactly where the errors occur. To achieve this the column index will need to be available to the frontend in the same way that the row index is, currently there's no simple way to map the column name where the error occurred to the index in the error data (see #199)

Also resolves #191 as it now shows an error summary when a required field is not used in the mapping step.