Add support for usage measurement where enabled
When a user of the prototype kit creates a new prototype, they are asked
if they would like to allow usage data to be shared with the authors.
This change uses that permission to measure usage unless the user has
explicitly disabled the plugin from collecting metrics, or changes the
prototype kit permission to false.