Add info when running prototype about stats
This commit adds an explanation to the console when running the
prototype kit, about the data importer's anonymous usage data.

Currently it is set to the following when the user has agreed to allow
usage data, and has not explicitly disabled it.

```
[WebServer] Watching dependency @register-dynamics/importer as it's a symbolic link.
[WebServer]
[WebServer] You are currently sending anonymous usage data to the prototype kit team.
[WebServer] In addition to this, anonymous usage data is also being sent to the authors
[WebServer] of the Data Upload Design Kit.
[WebServer] Should you wish to disable this you can do so by adding
[WebServer]
[WebServer]     disableImporterUsageData: true
[WebServer]
[WebServer] to the usage-data-config.json file in the prototype's directory.
```

We initially discussed doing this during the postinstall step, but this
results in it being output too early (before the user has accepted on
the first npm run dev) and obscured by more following content.