Auto-populate field names in the mapper if they match (#209)
If the name of a column in the user-submitted data matches the name
of a configured field, we now automatically select the field in the
column's mapping drop-down.

We only do this if the user previously hasn't submitted a mapping.
This is so that any fields they left blank will continue to be blank
and won't be overriden by the default if they come back to the mapping
page to correct an error.