Support secondary route for url mappers
When starting the process, we want to be able to provide a secondary
alternative route to the route lookup in template, so that we can skip
to the next page should configuration allow it.

This PR adds a 'sheetSelection' configuration property that can be set
to 'manual' or 'automatic'.  When set to automatic, and only a single
sheet is found in the upload, the sheet selection step is skipped.