Manage collection of usage data (#207)
* Add support for usage measurement where enabled

When a user of the prototype kit creates a new prototype, they are asked
if they would like to allow usage data to be shared with the authors.
This change uses that permission to measure usage unless the user has
explicitly disabled the plugin from collecting metrics, or changes the
prototype kit permission to false.

* Add info when running prototype about stats

This commit adds an explanation to the console when running the
prototype kit, about the data importer's anonymous usage data.

Currently it is set to the following when the user has agreed to allow
usage data, and has not explicitly disabled it.

```
[WebServer] Watching dependency @register-dynamics/importer as it's a symbolic link.
[WebServer]
[WebServer] You are currently sending anonymous usage data to the prototype kit team.
[WebServer] In addition to this, anonymous usage data is also being sent to the authors
[WebServer] of the Data Upload Design Kit.
[WebServer] Should you wish to disable this you can do so by adding
[WebServer]
[WebServer]     disableImporterUsageData: true
[WebServer]
[WebServer] to the usage-data-config.json file in the prototype's directory.
```

We initially discussed doing this during the postinstall step, but this
results in it being output too early (before the user has accepted on
the first npm run dev) and obscured by more following content.

* Refactor to use separate setting for usage data permissions

* Add config option to allow RD to collect usage data

* Fix dodgy acronym

* Clarify use of anonymous ID

* Add hint to contact us on github

* Fix incorrect use of new var name