Record usage events where permission granted (#227)
* Record usage events where permission granted

Where the user has possibly granted permission for us to record usage
data, we will now make a HTTPS request to an RD server, containing only
the client's uuid (generated by either us or the prototype kit depending
on permissions) and the name of the event. If the recording fails, then
we will not interupt the usage of the prototype at all, instead just
recording the failure in the user's terminal.

* Clarify comment