Add support for row numbers in table views
Row numbers can now be added to a table view by passing in a
`rowNumbers=N` property. The value of N is the row number of the
first row. Subsequent rows are just this number incremented.

This is almost certainly the wrong approach – the row number should
probably be supplier on a per-row basis. This would allow us to
display non-contiguous areas in the same table, for example.