Make plugin session creator safe for use as a copy constructor
Currently if we create a Session object in the frontend, it will always
create a new session and reset headers and ranges rather than load them
from the provided object (if present).

This now makes the constructor safe for use when reloading the express
session data.

In addition this commit moves the list of sheet names from a local
property to a function property that will query the backend for the list
of sheet names.