Restructure the code for comprehension (#248)
* Restructure the code for comprehension

* Make plugin session creator safe for use as a copy constructor

Currently if we create a Session object in the frontend, it will always
create a new session and reset headers and ranges rather than load them
from the provided object (if present).

This now makes the constructor safe for use when reloading the express
session data.

In addition this commit moves the list of sheet names from a local
property to a function property that will query the backend for the list
of sheet names.

* Remove the optional chaining check

* Move header/footer range storage to backend

* More move of session into backend

Moves the mapping data into backend, and now most methods on the session
object proxy their requests to the backend.  This is not intended to be
the final design, but a stepping stone on to a distinct backend library

* Make sure the MapData function has all the data it needs

* Make SessionSetFooterRange match SessionSetHeaderRange