Split out types (what the app wants to receive) and formats (variations on the mapping into that type, based on the shape of the input data).
This adds an extra argument to attribute-types.mapperForField, a format
specifier which is ignored except for (for now) date types. Even then
it's optional, as if left unspecified or false it'll return a date
mapper that only accepts spreadsheet date values and doesn't parse text.

So I hope that doesn't break the existing UI logic!

It also changes the return type of SessionGuessTypes to have the type list
for each column be a map of possible types, mapping type names to either
false (for types without formats) or an Array of possible format names.

The guesser will include "false" as a possible format the date type if the
input column contained spreadsheet date cells (that have no format). It
makes sense for false to coexist with other formats for a column containing
a mix of date cells and text cells that happen to contain dates.

The type and format names are in the format accepted by mapperForField.