Add dynamic session store (#261)
Adds support for different session storage mechanisms, initially moving the current implementation to a new class, and adding a new implementation which stores the session on disk. This latter mechanism is intended for development purposes so that it is possible to reload a page where state exists without having to re-run through the entire process, it is unlikely to be valuable when actually testing/using the prototype and so the default (MEMORY) is recommended.

The disk based storage required changes to other parts of the system as it can no longer rely on using references to the in-memory session, instead having to explicitly get/set (load and save) the session each time we want to write to it. To make this easier, an apply method was added to the session store class so that a function (potentially with a closure over some local vars) can be run against the session which is then saved.

In development (when there is no NODE_ENV set, or it is set to "development" the disk store is automatically enabled. When the prototype is closed (through either SIGINT or SIGTERM) then the JSON files containing the state will be removed from disk to avoid cluttering up the user's drive.