#753: Type guessing. (#260)
#753: Type guessing.

* We add a new function to the backend API, SessionGuessTypes. It's
similar in nature to SessionGetInputSampleRows, but for a given data range
it returns whether each column is "full" (contains no blank cells) and
(more interestingly) a list of possible types that match the values
found in that column.

The guts of it are in lib/importer/src/dudk/types/guesser.js

* We also add a SessionGetSupportedTypes function to the backend.

This just returns some static data indicating what types and formats
we support, so that any UI needing to let the user select a valid format
for a type can know what the options are, and get some human-readable names
to use.