Add row support to table view (#267)
Adds support for showing absolute row numbers for rows returned from the backend. These numbers are calculated in the backend and added to each row returned, changing the structure from an array of arrays to an array of objects where each object has two keys, index and row. The index is the absolute row index, the row is the same array as the original structure.

As part of this PR, we have updated how the tableView macro receives its parameters, moving to individual parameters to a paramaters object in line with the govuk design system. The keys expected by the arguments to tableView are:

    caption - A text caption for the table
    showHeaders - A boolean which determines whether headers are shown or not
    headers - The headers for this table
    showRowNumbers - A boolean which specifies whether to show row numbers in the left column or not
    rows - The row data containing the index of the row and the values for each cell
    moreRowsCount - How many more rows are available but now shown, useful when showing only n of many.

Other macros may wish to follow this pattern as they migrate to a more relevant namespace than the current one (importer).

A new filter function has been added to simplify debugging, the debug filter which makes it easier to discover the shape of data provided to the templates.