Error view for review page (#273)
Introduces an error view macro (importerErrorView) which shows users any errors of warnings encountered during the mapping process. The data representation for erros and warnings is now grouped by the error itself, showing the rows and fields where it occurs.

{
   message: "An error",
   meta: {
     first:  1,
     count: 1,
   },
  fields: ["A"],
  rows: []
}

rows in the structure above contains all of the rows containing errors along with the previous and following rows for extra context. The specific cells containing the error in each row has a new tag, error: true to show that the cell contains an error.