Use table view for range selection (#291)
The range_selector macro did not use the tableView and therefore there
was a duplication of code, and an error where an extra tbody was
included in tables with no header chrome.

This commit removes the range_selector in favour of the tableView, and
simplifies the header/footer selection step.