Use middleware to ensure errors on the page are cleaned up on next request (#292)
We use the session to hold error information to be displayed on the page. We previously used a function that could be called which would remove any error artifacts from the session before the request was processed. This was however not applied consistently.

This commit now introduces a middleware which will remove session data from the session at the start of each request.