Correctly map errors into presentation structure (#293)
The previous version of this function was incorrectly locating the cells with errors in based on the source headers (e.g. titles in sheet) instead of the mapping. This meant that we were not identifying the errors in each row, unless the mapping was identical to the source headers.

This change ensures that the errors are shown and are grouped correctly along with local context (previous and following rows).

Updating of errors into include field name will be added in a separate PR