Replace message in errors with error type/value/variant.
Rather than generate the error message in the backend, we know generate
an error structure instead which can be used to generate error messages.
This functionality should live in the core dudk code, but be separate
from the general backend functionality as it is very front-end specific
and will be configured from a client.