Copy GOV.UK change to use masks for footer image
In https://github.com/alphagov/govuk-frontend/commit/4c8d91acacf9d14dcc7e485fbf7e0f699ca441ad
GOV.UK Frontend moved to using the Royal Arms as a mask, which allows
it to copy the current colour of the text, which in theory is more
accessible.

So to get our own logo to display there, we need to copy that decision
by overriding the mask image in browsers that support it, and setting
the background on the ::before psudeo-element instead of the link
itself.