Don't include footer row in example values during mapping
Currently during mapping we identify the data range to select some
examples from, but the call that is made expects an inclusive range, and
we want an exclusive one. Due to the lack of options wrt range types we
instead just decrement the range we want to retrieve.

We may want to consider having different range types to avoid having to
add/subtract one based on requirements, and perhaps make
iteration and calculations easier