Allow skipping of header selection page (#306)
It's possible that a spreadsheet does not have headers, and so this PR makes it possible for a user to skip the header selection step by clicking the Skip button.

<img width="1310" height="723" alt="Upload_monthly_pension_return_–_GOV_UK_Prototype_Kit" src="https://github.com/user-attachments/assets/793c811f-2e69-4f4e-b723-df0e9787cf5b" />

Before this button existed, not selecting a header row achieved the same result, which was that the first row was chosen as headers.  This means that when, for instance, using tribble.xlsx we get the following result on the mapping page:

<img width="1043" height="463" alt="Upload_monthly_pension_return_–_GOV_UK_Prototype_Kit" src="https://github.com/user-attachments/assets/0e479383-faff-4878-8e1e-5a6065570048" />

This isn't ideal as the first row (the default when nothing selected) could contain useful data.  

To work around this assumption of a non-null header, we instead make the row for the header range out of bounds (e.g. -1) so that we can use it as a sentinel.  This allows us to fetch the column names during mapping as the base26 value of the index, which results in a mapping screen of 

<img width="1031" height="760" alt="Upload_monthly_pension_return_–_GOV_UK_Prototype_Kit" src="https://github.com/user-attachments/assets/51b7ee0e-1411-4bfc-9cf9-b054852daa0a" />

Unfortunately this approach requires that the backend expose the raw dimensions of the sheet so that the appropriate number of column headings can be generated.