E2E tests in CI (#307)
Fixes the E2E tests in CI so that they always run to completion.

We were used disk mode for the session store in tests, and it appears it was interfering with some of the tests - and suggests that we may have an issue with session cleanup (particularly the in-memory caches) which we can resolve in another PR.

There was a recent change to how examples are generated which broke the other test, but all should be working as expected now.