Add lib changes to support a new config parameter
Adds support to the core lib for a new config parameter which specifies
what do to when an upload has only one sheet.  'Manual' will, as now,
expect the user to select it (perhaps as a check that the correct file
was used), while 'automatic' will progress the user straight onto the
next page.

To support this behaviour, we have added support for an 'other' link
alongside the 'next' link when specifying form paths.