Provide sample files for trying out the prototype (#325)
The prototype requires users to have a file to upload, in order to try it out. To make this easier, we now provide two sample files (from the fixtures) which can be used by the user instead of having to build their own.

The files (pension data and tribbles) are included in the prototypes samples folder, and an endpoint was added to the demo to facilitate the download of these files, should the user wish to see them. The prototype itself has had changes to the index page giving the user a link to download, or a button to use that sample without needing to download it.

This functionality required changes in the plugin, which only exports an initialise function, using it to then set up multiple routes. The new route added is for using samples and requires a POST request with a parameter called samplefile. This parameter should be the name of a file in the samples folder of the prototype. Once submitted the user is directed to select a sheet as if they had manually uploaded a file.