Add version matrix to lint, codeql and test (#326)
As the prototype kit requires any nodejs version after 16, this PR introduces test matrices for codeql, linting and tests for versions from v16 onwards.

We don't lint v16 or v18 as they are not supported by our linter.
We don't test v16 as it is unsupported by our test framework.