Set current selection values on header/footer selection screens (#328)
When selecting header and footer rows, they are stored in the current session for when the user gets to the mapping screen. Unfortunately, if the user goes back to the header or footer selection screen, their previous selections are not shown even if they are still held in the backend state.

This PR includes the current header/footer selection in the appropriate hidden form elements that denote the previous selection. Because when showing footers the row index in the range is for the entire sheet, not just the n rows being shown, this was failing to highlight the footer rows (when present). To resolve this, we added data-row-index to the rows in the table, allowing us to access the row directly. Perhaps it might be worth adding data-cell-position to TDs or data-column-index to columns to avoid these kind of errors in future?